requirements.txt
streamlit>=1.30.0
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
plotly>=5.18.0
seaborn>=0.13.0
scipy>=1.11.0
exifread>=3.0.0
tf-keras-vis>=0.8.0
scikit-learn>=1.3.0
pandas>=2.1.0
